Add the view details ability to the credential issuer table as well. This required splitting this out into a separate component modal to make it reusable.
