+++ client/src/components/CredentialIssueCard.tsx
@@ -4,16 +4,11 @@ import {
   Typography,
   CardActions,
   Button,
-  Box,
-  Modal,
-  CircularProgress,
 } from "@mui/material";
 import { CredentialIssue } from "../utils/user.util";
-import { useEffect, useState } from "react";
-import { publicClient } from "../utils/client";
-import { institutionCredentialAbi } from "../utils/abi.util";
-import { getJsonDataFromPinata } from "../utils/pinata.util";
+import { useState } from "react";
 import { useUser } from "../context/UserContext";
+import { ViewCredentialDetailsModal } from "./ViewCredentialDetailsModal";
 
 export function CredentialIssueCard({
   credentialIssue,
@@ -21,41 +16,9 @@ export function CredentialIssueCard({
   credentialIssue: CredentialIssue;
 }) {
   const [viewMoreDetails, setViewMoreDetails] = useState(false);
-  const [moreDetails, setMoreDetails] = useState<JSON | object | null>(null);
-  const [loadingDetails, setLoadingDetails] = useState(true);
   const [hidden, setHidden] = useState(credentialIssue.hidden);
   const [saving, setSaving] = useState(false);
   const { fetchUserData } = useUser();
-
-  useEffect(() => {
-    const loadDetails = async () => {
-      let pinataJson: JSON | null = null;
-      if (credentialIssue.credential_type.issuer.contract_address) {
-        const cid = (await publicClient.readContract({
-          address: credentialIssue.credential_type.issuer.contract_address,
-          abi: institutionCredentialAbi,
-          functionName: "uri",
-          args: [
-            BigInt(credentialIssue.credential_type.token_id.replace(/n$/, "")),
-          ],
-        })) as string;
-
-        pinataJson = await getJsonDataFromPinata(cid);
-      }
-
-      setMoreDetails(pinataJson ?? {});
-      setLoadingDetails(false);
-    };
-
-    if (moreDetails === null && loadingDetails) {
-      loadDetails();
-    }
-  }, [
-    moreDetails,
-    loadingDetails,
-    credentialIssue.credential_type.issuer.contract_address,
-    credentialIssue.credential_type.token_id,
-  ]);
 
   const handleToggleHidden = async () => {
     setSaving(true);
@@ -88,60 +51,14 @@ export function CredentialIssueCard({
 
   return (
     <>
-      <Modal
-        open={viewMoreDetails}
-        onClose={() => setViewMoreDetails(false)}
-        aria-labelledby="modal-modal-title"
-        aria-describedby="modal-modal-description"
-      >
-        <Box
-          sx={{
-            display: "flex",
-            justifyContent: "center",
-            alignItems: "center",
-            height: "100vh",
-          }}
-          onClick={() => setViewMoreDetails(false)}
-        >
-          {moreDetails === null && loadingDetails ? (
-            <CircularProgress />
-          ) : (
-            <Card
-              sx={{
-                width: "100%",
-                maxWidth: 500,
-                bgcolor: "white",
-                boxShadow: 3,
-              }}
-            >
-              <CardContent>
-                <Typography variant="h6" gutterBottom>
-                  Additional Information
-                </Typography>
-
-                {moreDetails !== null && Object.keys(moreDetails).length > 0 ? (
-                  Object.entries(moreDetails).map(([key, value]) => (
-                    <Box key={key} sx={{ display: "flex", gap: 1, py: 0.5 }}>
-                      <Typography
-                        variant="body2"
-                        fontWeight="bold"
-                        sx={{ textTransform: "capitalize" }}
-                      >
-                        {key}:
-                      </Typography>
-                      <Typography variant="body2">{String(value)}</Typography>
-                    </Box>
-                  ))
-                ) : (
-                  <Typography variant="body2" color="text.secondary">
-                    There are no extra details from the institution.
-                  </Typography>
-                )}
-              </CardContent>
-            </Card>
-          )}
-        </Box>
-      </Modal>
+      <ViewCredentialDetailsModal
+        viewDetails={viewMoreDetails}
+        setViewDetails={setViewMoreDetails}
+        tokenId={credentialIssue.credential_type.token_id}
+        contractAddress={
+          credentialIssue.credential_type.issuer.contract_address
+        }
+      />
       <Card sx={{ minWidth: 375, display: "flex", flexDirection: "column" }}>
         <CardContent sx={{ flexGrow: 1 }}>
           <Typography gutterBottom sx={{ color: "white" }}>
